requirements.txt
chromadb>=0.5.5
sentence-transformers>=3.0.1
google-genai>=1.0.0
PyYAML>=6.0.1
python-dateutil>=2.9.0.post0
numpy>=1.26.0
